+++ android/src/main/java/io/customer/reactnative/sdk/CustomerIOReactNativeModule.kt
@@ -1,16 +1,16 @@
 package io.customer.reactnative.sdk
 
-import com.facebook.react.bridge.ReactApplicationContext
-import com.facebook.react.bridge.ReactContextBaseJavaModule
-import com.facebook.react.bridge.ReactMethod
-import com.facebook.react.bridge.ReadableMap
+import com.facebook.react.bridge.*
 import io.customer.reactnative.sdk.extension.toMap
+import io.customer.reactnative.sdk.messagingpush.RNCIOPushMessaging
 import io.customer.sdk.CustomerIO
 import io.customer.sdk.CustomerIOShared
 import io.customer.sdk.util.Logger
 
 class CustomerIOReactNativeModule(
     reactContext: ReactApplicationContext,
+    private val pushMessagingModule: RNCIOPushMessaging,
+    private val inAppMessagingModule: RNCIOInAppMessaging,
 ) : ReactContextBaseJavaModule(reactContext) {
     private val logger: Logger
         get() = CustomerIOShared.instance().diStaticGraph.logger
@@ -53,6 +53,7 @@ class CustomerIOReactNativeModule(
                 environment = env,
                 configuration = config,
                 packageConfig = packageConfig,
+                inAppEventListener = inAppMessagingModule,
             )
             logger.info("Customer.io instance initialized successfully from app")
         } catch (ex: Exception) {
@@ -100,6 +101,23 @@ class CustomerIOReactNativeModule(
         if (isNotInitialized()) return
 
         customerIO.screen(name, attributes.toMap())
+    }
+
+    @ReactMethod
+    fun registerDeviceToken(token: String) {
+        if (isNotInitialized()) return
+
+        customerIO.registerDeviceToken(token)
+    }
+
+    @ReactMethod
+    fun getPushPermissionStatus(promise: Promise) {
+        pushMessagingModule.getPushPermissionStatus(promise)
+    }
+
+    @ReactMethod
+    fun showPromptForPushNotifications(pushConfigurationOptions: ReadableMap?, promise: Promise) {
+        pushMessagingModule.showPromptForPushNotifications(pushConfigurationOptions, promise)
     }
 
     companion object {