+++ android/src/main/java/io/customer/reactnative/sdk/CustomerIOReactNativePackage.kt
@@ -4,11 +4,20 @@ import com.facebook.react.ReactPackage
 import com.facebook.react.bridge.NativeModule
 import com.facebook.react.bridge.ReactApplicationContext
 import com.facebook.react.uimanager.ViewManager
+import io.customer.reactnative.sdk.messagingpush.RNCIOPushMessaging
 
 class CustomerIOReactNativePackage : ReactPackage {
     override fun createNativeModules(reactContext: ReactApplicationContext): List<NativeModule> {
+        val pushMessagingModule = RNCIOPushMessaging(reactContext)
+        val inAppMessagingModule = RNCIOInAppMessaging(reactContext)
         return listOf(
-            CustomerIOReactNativeModule(reactContext),
+            inAppMessagingModule,
+            pushMessagingModule,
+            CustomerIOReactNativeModule(
+                reactContext = reactContext,
+                pushMessagingModule = pushMessagingModule,
+                inAppMessagingModule = inAppMessagingModule,
+            ),
         )
     }