+++ ios/CustomerioReactnative.m
@@ -21,4 +21,12 @@ RCT_EXTERN_METHOD(setProfileAttributes : (nonnull NSDictionary *) data)
 RCT_EXTERN_METHOD(screen: (nonnull NSString *) name
                   data : (NSDictionary *) data)
 
+RCT_EXTERN_METHOD(registerDeviceToken : (nonnull NSString *) token)
+
+RCT_EXTERN_METHOD(showPromptForPushNotifications: (NSDictionary *) options
+                  resolver:(RCTPromiseResolveBlock)resolve
+                  rejecter:(RCTPromiseRejectBlock)reject)
+
+RCT_EXTERN_METHOD(getPushPermissionStatus: (RCTPromiseResolveBlock) resolver
+                  rejecter:(RCTPromiseRejectBlock)rejecter)
 @end