+++ src/types/index.tsx
@@ -1,0 +1,2 @@
+export * from './PushPermissionOptions';
+export * from './PushPermissionStatus';